+++ internal/table/client.go
@@ -42,6 +42,7 @@ func New(ctx context.Context, cc grpc.ClientConnInterface, config *config.Config
 		pool: pool.New[*Session, Session](ctx,
 			pool.WithLimit[*Session, Session](config.SizeLimit()),
 			pool.WithItemUsageLimit[*Session, Session](config.SessionUsageLimit()),
+			pool.WithItemUsageTTL[*Session, Session](config.SessionUsageTTL()),
 			pool.WithIdleTimeToLive[*Session, Session](config.IdleThreshold()),
 			pool.WithCreateItemTimeout[*Session, Session](config.CreateSessionTimeout()),
 			pool.WithCloseItemTimeout[*Session, Session](config.DeleteTimeout()),